TST: test_tools_matrix.py reorganization.
